feat: implement modules rewriting
